+++ bridges/snowbridge/pallets/inbound-queue-v2/src/mock.rs
@@ -24,6 +24,7 @@ use sp_runtime::{
 use sp_std::{convert::From, default::Default, marker::PhantomData};
 use xcm::{latest::SendXcm, opaque::latest::WESTEND_GENESIS_HASH, prelude::*};
 type Block = frame_system::mocking::MockBlock<Test>;
+pub use snowbridge_test_utils::mock_xcm::{MockXcmExecutor, MockXcmSender, Weightless};
 
 frame_support::construct_runtime!(
 	pub enum Test
@@ -111,52 +112,6 @@ const GATEWAY_ADDRESS: [u8; 20] = hex!["b1185ede04202fe62d38f5db72f71e38ff3e8305
 impl<T: snowbridge_pallet_ethereum_client::Config> BenchmarkHelper<T> for Test {
 	// not implemented since the MockVerifier is used for tests
 	fn initialize_storage(_: BeaconHeader, _: H256) {}
-}
-
-// Mock XCM sender that always succeeds
-pub struct MockXcmSender;
-impl SendXcm for MockXcmSender {
-	type Ticket = Xcm<()>;
-
-	fn validate(
-		dest: &mut Option<Location>,
-		xcm: &mut Option<Xcm<()>>,
-	) -> SendResult<Self::Ticket> {
-		if let Some(location) = dest {
-			match location.unpack() {
-				(_, [Parachain(1001)]) => return Err(SendError::NotApplicable),
-				_ => Ok((xcm.clone().unwrap(), Assets::default())),
-			}
-		} else {
-			Ok((xcm.clone().unwrap(), Assets::default()))
-		}
-	}
-
-	fn deliver(xcm: Self::Ticket) -> core::result::Result<XcmHash, SendError> {
-		let hash = xcm.using_encoded(sp_io::hashing::blake2_256);
-		Ok(hash)
-	}
-}
-
-pub enum Weightless {}
-impl PreparedMessage for Weightless {
-	fn weight_of(&self) -> Weight {
-		unreachable!();
-	}
-}
-
-pub struct MockXcmExecutor;
-impl<C> ExecuteXcm<C> for MockXcmExecutor {
-	type Prepared = Weightless;
-	fn prepare(message: Xcm<C>) -> Result<Self::Prepared, Xcm<C>> {
-		Err(message)
-	}
-	fn execute(_: impl Into<Location>, _: Self::Prepared, _: &mut XcmHash, _: Weight) -> Outcome {
-		unreachable!()
-	}
-	fn charge_fees(_: impl Into<Location>, _: Assets) -> xcm::latest::Result {
-		Ok(())
-	}
 }
 
 pub struct MockTokenIdConvert;
@@ -323,297 +278,5 @@ impl WeightInfo for () {
 			.saturating_add(Weight::from_parts(0, 3601))
 			.saturating_add(RocksDbWeight::get().reads(2))
 			.saturating_add(RocksDbWeight::get().writes(2))
-	}
-}
-
-pub mod mock_xcm_send_failure {
-	use super::*;
-
-	frame_support::construct_runtime!(
-		pub enum TestXcmSendFailure
-		{
-			System: frame_system::{Pallet, Call, Storage, Event<T>},
-			Balances: pallet_balances::{Pallet, Call, Storage, Config<T>, Event<T>},
-			EthereumBeaconClient: snowbridge_pallet_ethereum_client::{Pallet, Call, Storage, Event<T>},
-			InboundQueue: inbound_queue_v2::{Pallet, Call, Storage, Event<T>},
-		}
-	);
-
-	#[derive_impl(frame_system::config_preludes::TestDefaultConfig)]
-	impl frame_system::Config for TestXcmSendFailure {
-		type AccountId = AccountId;
-		type Lookup = IdentityLookup<Self::AccountId>;
-		type AccountData = pallet_balances::AccountData<u128>;
-		type Block = Block;
-	}
-
-	#[derive_impl(pallet_balances::config_preludes::TestDefaultConfig)]
-	impl pallet_balances::Config for TestXcmSendFailure {
-		type Balance = Balance;
-		type ExistentialDeposit = ExistentialDeposit;
-		type AccountStore = System;
-	}
-
-	impl inbound_queue_v2::Config for TestXcmSendFailure {
-		type RuntimeEvent = RuntimeEvent;
-		type Verifier = MockVerifier;
-		type XcmSender = MockXcmFailureSender;
-		type XcmExecutor = MockXcmExecutor;
-		type RewardPayment = ();
-		type EthereumNetwork = EthereumNetwork;
-		type GatewayAddress = GatewayAddress;
-		type AssetHubParaId = ConstU32<1000>;
-		type MessageConverter = MessageToXcm<
-			CreateAssetCall,
-			CreateAssetDeposit,
-			EthereumNetwork,
-			InboundQueueLocation,
-			MockTokenIdConvert,
-			GatewayAddress,
-			UniversalLocation,
-			AssetHubFromEthereum,
-		>;
-		#[cfg(feature = "runtime-benchmarks")]
-		type Helper = Test;
-		type Balance = u128;
-		type WeightInfo = ();
-		type WeightToFee = IdentityFee<u128>;
-		type Token = Balances;
-		type AccountToLocation = MockAccountLocationConverter<AccountId>;
-		type RewardKind = BridgeReward;
-		type DefaultRewardKind = DefaultMyRewardKind;
-	}
-
-	impl snowbridge_pallet_ethereum_client::Config for TestXcmSendFailure {
-		type RuntimeEvent = RuntimeEvent;
-		type ForkVersions = ChainForkVersions;
-		type FreeHeadersInterval = ConstU32<32>;
-		type WeightInfo = ();
-	}
-
-	pub struct MockXcmFailureSender;
-	impl SendXcm for MockXcmFailureSender {
-		type Ticket = Xcm<()>;
-
-		fn validate(
-			dest: &mut Option<Location>,
-			xcm: &mut Option<Xcm<()>>,
-		) -> SendResult<Self::Ticket> {
-			if let Some(location) = dest {
-				match location.unpack() {
-					(_, [Parachain(1001)]) => return Err(SendError::NotApplicable),
-					_ => Ok((xcm.clone().unwrap(), Assets::default())),
-				}
-			} else {
-				Ok((xcm.clone().unwrap(), Assets::default()))
-			}
-		}
-
-		fn deliver(_xcm: Self::Ticket) -> core::result::Result<XcmHash, SendError> {
-			return Err(SendError::DestinationUnsupported)
-		}
-	}
-
-	pub fn new_tester() -> sp_io::TestExternalities {
-		let storage = frame_system::GenesisConfig::<TestXcmSendFailure>::default()
-			.build_storage()
-			.unwrap();
-		let mut ext: sp_io::TestExternalities = storage.into();
-		ext.execute_with(setup);
-		ext
-	}
-}
-
-pub mod mock_xcm_validate_failure {
-	use super::*;
-
-	#[cfg(feature = "runtime-benchmarks")]
-	impl<T: snowbridge_pallet_ethereum_client::Config> BenchmarkHelper<T> for Test {
-		// not implemented since the MockVerifier is used for tests
-		fn initialize_storage(_: BeaconHeader, _: H256) {}
-	}
-
-	frame_support::construct_runtime!(
-		pub enum Test
-		{
-			System: frame_system::{Pallet, Call, Storage, Event<T>},
-			Balances: pallet_balances::{Pallet, Call, Storage, Config<T>, Event<T>},
-			EthereumBeaconClient: snowbridge_pallet_ethereum_client::{Pallet, Call, Storage, Event<T>},
-			InboundQueue: inbound_queue_v2::{Pallet, Call, Storage, Event<T>},
-		}
-	);
-
-	#[derive_impl(frame_system::config_preludes::TestDefaultConfig)]
-	impl frame_system::Config for Test {
-		type AccountId = AccountId;
-		type Lookup = IdentityLookup<Self::AccountId>;
-		type AccountData = pallet_balances::AccountData<u128>;
-		type Block = Block;
-	}
-
-	#[derive_impl(pallet_balances::config_preludes::TestDefaultConfig)]
-	impl pallet_balances::Config for Test {
-		type Balance = Balance;
-		type ExistentialDeposit = ExistentialDeposit;
-		type AccountStore = System;
-	}
-
-	impl inbound_queue_v2::Config for Test {
-		type RuntimeEvent = RuntimeEvent;
-		type Verifier = MockVerifier;
-		type XcmSender = MockXcmFailureValidate;
-		type XcmExecutor = MockXcmExecutor;
-		type RewardPayment = ();
-		type EthereumNetwork = EthereumNetwork;
-		type GatewayAddress = GatewayAddress;
-		type AssetHubParaId = ConstU32<1000>;
-		type MessageConverter = MessageToXcm<
-			CreateAssetCall,
-			CreateAssetDeposit,
-			EthereumNetwork,
-			InboundQueueLocation,
-			MockTokenIdConvert,
-			GatewayAddress,
-			UniversalLocation,
-			AssetHubFromEthereum,
-		>;
-		#[cfg(feature = "runtime-benchmarks")]
-		type Helper = Test;
-		type Balance = u128;
-		type WeightInfo = ();
-		type WeightToFee = IdentityFee<u128>;
-		type Token = Balances;
-		type AccountToLocation = MockAccountLocationConverter<AccountId>;
-		type RewardKind = BridgeReward;
-		type DefaultRewardKind = DefaultMyRewardKind;
-	}
-
-	impl snowbridge_pallet_ethereum_client::Config for Test {
-		type RuntimeEvent = RuntimeEvent;
-		type ForkVersions = ChainForkVersions;
-		type FreeHeadersInterval = ConstU32<32>;
-		type WeightInfo = ();
-	}
-
-	pub struct MockXcmFailureValidate;
-	impl SendXcm for MockXcmFailureValidate {
-		type Ticket = Xcm<()>;
-
-		fn validate(
-			_dest: &mut Option<Location>,
-			_xcm: &mut Option<Xcm<()>>,
-		) -> SendResult<Self::Ticket> {
-			return Err(SendError::NotApplicable)
-		}
-
-		fn deliver(xcm: Self::Ticket) -> core::result::Result<XcmHash, SendError> {
-			let hash = xcm.using_encoded(sp_io::hashing::blake2_256);
-			Ok(hash)
-		}
-	}
-
-	pub fn new_tester() -> sp_io::TestExternalities {
-		let storage = frame_system::GenesisConfig::<Test>::default().build_storage().unwrap();
-		let mut ext: sp_io::TestExternalities = storage.into();
-		ext.execute_with(setup);
-		ext
-	}
-}
-
-pub mod mock_charge_fees_failure {
-	use super::*;
-
-	#[cfg(feature = "runtime-benchmarks")]
-	impl<T: snowbridge_pallet_ethereum_client::Config> BenchmarkHelper<T> for Test {
-		// not implemented since the MockVerifier is used for tests
-		fn initialize_storage(_: BeaconHeader, _: H256) {}
-	}
-
-	frame_support::construct_runtime!(
-		pub enum Test
-		{
-			System: frame_system::{Pallet, Call, Storage, Event<T>},
-			Balances: pallet_balances::{Pallet, Call, Storage, Config<T>, Event<T>},
-			EthereumBeaconClient: snowbridge_pallet_ethereum_client::{Pallet, Call, Storage, Event<T>},
-			InboundQueue: inbound_queue_v2::{Pallet, Call, Storage, Event<T>},
-		}
-	);
-
-	#[derive_impl(frame_system::config_preludes::TestDefaultConfig)]
-	impl frame_system::Config for Test {
-		type AccountId = AccountId;
-		type Lookup = IdentityLookup<Self::AccountId>;
-		type AccountData = pallet_balances::AccountData<u128>;
-		type Block = Block;
-	}
-
-	#[derive_impl(pallet_balances::config_preludes::TestDefaultConfig)]
-	impl pallet_balances::Config for Test {
-		type Balance = Balance;
-		type ExistentialDeposit = ExistentialDeposit;
-		type AccountStore = System;
-	}
-
-	impl inbound_queue_v2::Config for Test {
-		type RuntimeEvent = RuntimeEvent;
-		type Verifier = MockVerifier;
-		type XcmSender = MockXcmSender;
-		type XcmExecutor = MockXcmChargeFeesFailure;
-		type RewardPayment = ();
-		type EthereumNetwork = EthereumNetwork;
-		type GatewayAddress = GatewayAddress;
-		type AssetHubParaId = ConstU32<1000>;
-		type MessageConverter = MessageToXcm<
-			CreateAssetCall,
-			CreateAssetDeposit,
-			EthereumNetwork,
-			InboundQueueLocation,
-			MockTokenIdConvert,
-			GatewayAddress,
-			UniversalLocation,
-			AssetHubFromEthereum,
-		>;
-		#[cfg(feature = "runtime-benchmarks")]
-		type Helper = Test;
-		type Balance = u128;
-		type WeightInfo = ();
-		type WeightToFee = IdentityFee<u128>;
-		type Token = Balances;
-		type AccountToLocation = MockAccountLocationConverter<AccountId>;
-		type RewardKind = BridgeReward;
-		type DefaultRewardKind = DefaultMyRewardKind;
-	}
-
-	impl snowbridge_pallet_ethereum_client::Config for Test {
-		type RuntimeEvent = RuntimeEvent;
-		type ForkVersions = ChainForkVersions;
-		type FreeHeadersInterval = ConstU32<32>;
-		type WeightInfo = ();
-	}
-
-	pub struct MockXcmChargeFeesFailure;
-	impl<C> ExecuteXcm<C> for MockXcmChargeFeesFailure {
-		type Prepared = Weightless;
-		fn prepare(message: Xcm<C>) -> Result<Self::Prepared, Xcm<C>> {
-			Err(message)
-		}
-		fn execute(
-			_: impl Into<Location>,
-			_: Self::Prepared,
-			_: &mut XcmHash,
-			_: Weight,
-		) -> Outcome {
-			unreachable!()
-		}
-		fn charge_fees(_: impl Into<Location>, _: Assets) -> xcm::latest::Result {
-			Err(XcmError::Barrier)
-		}
-	}
-
-	pub fn new_tester() -> sp_io::TestExternalities {
-		let storage = frame_system::GenesisConfig::<Test>::default().build_storage().unwrap();
-		let mut ext: sp_io::TestExternalities = storage.into();
-		ext.execute_with(setup);
-		ext
 	}
 }
